--- cognite/powerops/client/data_classes.py
@@ -1,5 +1,0 @@
-# TODO: do we need these imports here? ruff wants to remove the imports
-# ruff: noqa
-
-from ._generated.cogshop1 import data_classes as cogshop1
-from .shop.shop_run import SHOPRun, SHOPRunList, SHOPRunStatus
