+++ cognite/powerops/client/shop/data_classes/__init__.py
@@ -1,16 +1,14 @@
-from cognite.powerops.client.shop.data_classes.case import Case
 from cognite.powerops.client.shop.data_classes.shop_result_files import ShopLogFile, ShopResultFile, ShopYamlFile
 from cognite.powerops.client.shop.data_classes.shop_results import ObjectiveFunction, ShopRunResult
-from cognite.powerops.client.shop.data_classes.shop_run import ShopRun
+from cognite.powerops.client.shop.data_classes.shop_run import SHOPRun
 from cognite.powerops.client.shop.data_classes.shop_run_event import ShopRunEvent
 
 __all__ = [
-    "Case",
     "ObjectiveFunction",
     "ShopRunResult",
     "ShopLogFile",
     "ShopResultFile",
-    "ShopRun",
+    "SHOPRun",
     "ShopRunEvent",
     "ShopYamlFile",
 ]
