+++ cognite/powerops/utils/identifiers.py
