+++ tests/test_unit/client/shop/test_shop_case/test_shop_case_loading.py
@@ -3,7 +3,7 @@ from pathlib import Path
 import pytest
 import yaml
 
-from cognite.powerops.client.shop.shop_case import SHOPCase
+from cognite.powerops.client.shop.data_classes.shop_case import SHOPCase
 
 
 @pytest.fixture
